% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropEmpty_row.R
\name{dropEmpty_row}
\alias{dropEmpty_row}
\title{Drop the empty rows of a data.frame}
\usage{
dropEmpty_row(input_data)
}
\arguments{
\item{input_data}{A data.frame.}
}
\value{
A data.frame where at least one variable was observed per row.
}
\description{
This function drops the rows of a data.frame that are full of NAs.
}
\examples{
dropEmpty_row(mtcars)

}
\author{
Stephen Semick
}

