% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code.R
\name{plot_code}
\alias{plot_code}
\title{Get the code for reproducing a plot}
\usage{
plot_code(filename, fulldata, selection = seq_len(nrow(fulldata)), x, y,
  color = "lightblue", pal = "Set1")
}
\arguments{
\item{filename}{The filename (preferably the full path) for the data
used.}

\item{fulldata}{The data.frame with the original full data displayed in
\link{explore}.}

\item{selection}{An integer vector of the rows of the \code{fulldata}
without empty rows (see \link{dropEmpty_row}) to be used in the plot.}

\item{x}{A variable name in \code{fulldata}.}

\item{y}{A variable name in \code{fulldata}. If missing, the code for a 
one way plot is shown.}

\item{color}{A color to use in the plots that have one single color.}

\item{pal}{A color palette to use with \link[RColorBrewer]{brewer.pal} for
the \link[vcd]{mosaic} plots when `x` and `y` are factors.}
}
\value{
A character vector with the code to reproduce a figure.
}
\description{
This function creates the code in text needed for reproducing a plot
from \link{explore}.
}
\examples{
cat(plot_code('mtcars.csv', mtcars, x = 'mpg', selection = 1:5))
cat(plot_code('mtcars.csv', mtcars, x = 'mpg', y = 'vs'))
}
\author{
Leonardo Collado-Torres
}

