% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oneway.R
\name{plot_oneway}
\alias{plot_oneway}
\title{Univariate plot}
\usage{
plot_oneway(info, title, color = "lightblue")
}
\arguments{
\item{info}{A vector.}

\item{title}{The title to use in the plot.}

\item{color}{A color to use in the plot.}
}
\value{
A plot.
}
\description{
Tries to guess the best plot to show for a single variable.
}
\examples{
plot_oneway(mtcars$mpg, "mpg")
plot_oneway(as.factor(mtcars$gear), "gear")

}
\author{
Leonardo Collado-Torres
}
