% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Univariate plot}
\usage{
plot_summary(info, title, color = "lightblue")
}
\arguments{
\item{info}{A vector.}

\item{title}{The title to use in the plot.}

\item{color}{A color to use in the plot.}
}
\value{
A plot.
}
\description{
Tries to guess the best plot to show for a single variable.
}
\examples{
plot_summary(mtcars$mpg, 'mpg')
plot_summary(as.factor(mtcars$gear), 'gear')

}
\author{
Leonardo Collado-Torres
}

