% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twoway.R
\name{plot_twoway}
\alias{plot_twoway}
\title{Plot two variables against each other}
\usage{
plot_twoway(x, y, xvar, yvar, color = "lightblue", pal = "Set1")
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector of the same length as `x`.}

\item{xvar}{The name of the x variable.}

\item{yvar}{The name of the y variable.}

\item{color}{A color to use in the plots that have one single color.}

\item{pal}{A color palette to use with \link[RColorBrewer]{brewer.pal} for
the \link[vcd]{mosaic} plots when `x` and `y` are factors.}
}
\value{
A plot of `x` versus `y`.
}
\description{
This function tries to choose a good plot to show for comparing two
variables at a time.
}
\examples{
plot_twoway(as.factor(mtcars$gear), mtcars$mpg, xvar = "gear", yvar = "mpg")

}
\author{
Leonardo Collado-Torres
}
