% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary.R
\name{stat_summary}
\alias{stat_summary}
\title{Statistical summary for one variable}
\usage{
stat_summary(info)
}
\arguments{
\item{info}{A numeric vector.}
}
\value{
The summary if `info` is numeric.
}
\description{
Print the summary for a numeric variable that includes the SD and the
variance.
}
\examples{
stat_summary(mtcars$mpg)

}
\author{
Leonardo Collado-Torres
}
