% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summary.R
\name{table_summary}
\alias{table_summary}
\title{Simple univariate table summary}
\usage{
table_summary(information)
}
\arguments{
\item{information}{A vector.}
}
\value{
A univariate table displaying the NAs if present. If `information` is a
character vector, it will split it by semicolons and show the table of the
resulting vector.
}
\description{
Prints a simple univariate table summary
}
\examples{
table_summary(mtcars$mpg)

}
\author{
Leonardo Collado-Torres
}
