% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore a CSV file or a data.frame stored in a Rdata file}
\usage{
explore(...)
}
\arguments{
\item{...}{Additional arguments passed to \link[shiny]{runApp}.}
}
\value{
The Shiny application is opened in your browser. To stop the
function from running you have to cancel the command using ctrl + c or the
equivalent shortcut in your system.
}
\description{
Opens the shiny application that allows you to explore a CSV file or a
data.frame stored in a Rdata file.
}
\details{
This function runs the Shiny application including in this package. It is basically a wrapper for \link[shiny]{runApp}.
}
\examples{
\dontrun{
## Open the Shiny application
explore()
## Remember to cancel the command to get back to the R console once you are done analyzing your own data.
}
}
\references{
For more information on what a Shiny app is check http://shiny.rstudio.com/
}
\author{
Stephen Semick, Leonardo Collado-Torres
}
