% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table}
\alias{read_table}
\title{Reads a table and does some minor formatting}
\usage{
read_table(path, name = path)
}
\arguments{
\item{path}{the path to a single file.}

\item{name}{the file name including the extension. Might be different
from `path` in the context of \link[shiny]{fileInput}.}
}
\value{
A data.frame where the POSIXct and POSIXt variables get
changed to Date variables.
}
\description{
This function reads a table and does some minor formatting on it. It
basically used \link[rio]{import} but can also handle `.rda` files.
}
\examples{

library("rio")
export(iris, "iris1.csv")
df <- read_table("iris1.csv")

}
\author{
Leonardo Collado-Torres
}
